+++ apps/prairielearn/src/lib/chunks.ts
@@ -16,12 +16,12 @@ import * as sqldb from '@prairielearn/postgres';
 
 import { getLockNameForCoursePath } from '../models/course.js';
 import * as courseDB from '../sync/course-db.js';
-import { CourseData } from '../sync/course-db.js';
+import { type CourseData } from '../sync/course-db.js';
 
 import { downloadFromS3, makeS3ClientConfig } from './aws.js';
 import { chalk, chalkDim } from './chalk.js';
 import { config } from './config.js';
-import { createServerJob, ServerJob } from './server-jobs.js';
+import { createServerJob, type ServerJob } from './server-jobs.js';
 
 const sql = sqldb.loadSqlEquiv(import.meta.url);