+++ apps/prairielearn/src/pages/studentAssessmentInstance/studentAssessmentInstance.html.ts
@@ -16,15 +16,19 @@ import {
   ExamQuestionAvailablePoints,
   ExamQuestionStatus,
   InstanceQuestionPoints,
-  QuestionAwardedPoints,
+  QuestionVariantHistory,
 } from '../../components/QuestionScore.html.js';
 import { Scorebar } from '../../components/Scorebar.html.js';
 import { StudentAccessRulesPopover } from '../../components/StudentAccessRulesPopover.html.js';
 import { TimeLimitExpiredModal } from '../../components/TimeLimitExpiredModal.html.js';
 import { compiledScriptTag } from '../../lib/assets.js';
-import { AssessmentInstance, GroupConfig, InstanceQuestion } from '../../lib/db-types.js';
+import {
+  type AssessmentInstance,
+  type GroupConfig,
+  type InstanceQuestion,
+} from '../../lib/db-types.js';
 import { formatPoints } from '../../lib/format.js';
-import { GroupInfo } from '../../lib/groups.js';
+import { type GroupInfo } from '../../lib/groups.js';
 
 export function StudentAssessmentInstance({
   showTimeLimitExpiredModal,
@@ -45,6 +49,22 @@ export function StudentAssessmentInstance({
     }
   | { groupConfig?: undefined; groupInfo?: undefined; userCanAssignRoles?: undefined }
 )) {
+  let savedAnswers = 0;
+  let suspendedSavedAnswers = 0;
+  resLocals.instance_questions.forEach((question) => {
+    if (question.status === 'saved') {
+      if (question.allow_grade_left_ms > 0) {
+        suspendedSavedAnswers++;
+      } else if (question.max_auto_points || !question.max_manual_points) {
+        // Note that we exclude questions that are not auto-graded from the count.
+        // This count is used to determine whether the "Grade X saved answers"
+        // button should be enabled, and clicking that button won't do any good for
+        // manually-graded questions.
+        savedAnswers++;
+      }
+    }
+  });
+
   // Keep this in sync with the `InstanceQuestionTableHeader` function below.
   const zoneTitleColspan = run(() => {
     const trailingColumnsCount =
@@ -228,19 +248,21 @@ export function StudentAssessmentInstance({
                       ? html`
                           <tr>
                             <th colspan="${zoneTitleColspan}">
-                              <span class="mr-1">${instance_question.zone_title}</span>
-                              ${instance_question.zone_has_max_points
-                                ? ZoneInfoBadge({
-                                    popoverContent: `Of the points that you are awarded for answering these questions, at most ${instance_question.zone_max_points} will count toward your total points.`,
-                                    mainContent: `Maximum ${instance_question.zone_max_points} points`,
-                                  })
-                                : ''}
-                              ${instance_question.zone_has_best_questions
-                                ? ZoneInfoBadge({
-                                    popoverContent: `Of these questions, only the ${instance_question.zone_best_questions} with the highest number of awarded points will count toward your total points.`,
-                                    mainContent: `Best ${instance_question.zone_best_questions} questions`,
-                                  })
-                                : ''}
+                              <div class="d-flex align-items-center">
+                                <span class="mr-2">${instance_question.zone_title}</span>
+                                ${instance_question.zone_has_max_points
+                                  ? ZoneInfoPopover({
+                                      label: `Maximum ${instance_question.zone_max_points} points`,
+                                      content: `Of the points that you are awarded for answering these questions, at most ${instance_question.zone_max_points} will count toward your total points.`,
+                                    })
+                                  : ''}
+                                ${instance_question.zone_has_best_questions
+                                  ? ZoneInfoPopover({
+                                      label: `Best ${instance_question.zone_best_questions} questions`,
+                                      content: `Of these questions, only the ${instance_question.zone_best_questions} with the highest number of awarded points will count toward your total points.`,
+                                    })
+                                  : ''}
+                              </div>
                             </th>
                           </tr>
                         `
@@ -264,26 +286,31 @@ export function StudentAssessmentInstance({
                       ${resLocals.assessment.type === 'Exam'
                         ? html`
                             <td class="text-center">
-                              ${ExamQuestionStatus({ instance_question })}
+                              ${ExamQuestionStatus({
+                                instance_question,
+                                assessment_question: instance_question, // Required fields are in instance_question
+                              })}
                             </td>
                             ${resLocals.has_auto_grading_question &&
                             resLocals.assessment.allow_real_time_grading
                               ? html`
                                   <td class="text-center">
-                                    ${ExamQuestionAvailablePoints({
-                                      open:
-                                        resLocals.assessment_instance.open &&
-                                        instance_question.open,
-                                      currentWeight:
-                                        instance_question.points_list_original[
-                                          instance_question.number_attempts
-                                        ] - instance_question.max_manual_points,
-                                      pointsList: instance_question.points_list.map(
-                                        (p) => p - instance_question.max_manual_points,
-                                      ),
-                                      highestSubmissionScore:
-                                        instance_question.highest_submission_score,
-                                    })}
+                                    ${instance_question.max_auto_points
+                                      ? ExamQuestionAvailablePoints({
+                                          open:
+                                            resLocals.assessment_instance.open &&
+                                            instance_question.open,
+                                          currentWeight:
+                                            instance_question.points_list_original[
+                                              instance_question.number_attempts
+                                            ] - instance_question.max_manual_points,
+                                          pointsList: instance_question.points_list.map(
+                                            (p) => p - instance_question.max_manual_points,
+                                          ),
+                                          highestSubmissionScore:
+                                            instance_question.highest_submission_score,
+                                        })
+                                      : html`&mdash;`}
                                   </td>
                                 `
                               : ''}
@@ -338,12 +365,22 @@ export function StudentAssessmentInstance({
                             ${resLocals.has_auto_grading_question
                               ? html`
                                   <td class="text-center">
-                                    <span class="badge badge-primary">
-                                      ${formatPoints(instance_question.current_value)}
-                                    </span>
+                                    ${run(() => {
+                                      if (!instance_question.max_auto_points) return html`&mdash;`;
+
+                                      // Compute the current "auto" value by subtracting the manual points.
+                                      // We use this because `current_value` doesn't account for manual points.
+                                      // We don't want to mislead the student into thinking that they can earn
+                                      // more points than they actually can.
+                                      const currentAutoValue =
+                                        (instance_question.current_value ?? 0) -
+                                        (instance_question.max_manual_points ?? 0);
+
+                                      return formatPoints(currentAutoValue);
+                                    })}
                                   </td>
                                   <td class="text-center">
-                                    ${QuestionAwardedPoints({
+                                    ${QuestionVariantHistory({
                                       urlPrefix: resLocals.urlPrefix,
                                       instanceQuestionId: instance_question.id,
                                       previousVariants: instance_question.previous_variants,
@@ -398,15 +435,15 @@ export function StudentAssessmentInstance({
                               name="__csrf_token"
                               value="${resLocals.__csrf_token}"
                             />
-                            ${resLocals.savedAnswers > 0
+                            ${savedAnswers > 0
                               ? html`
                                   <button
                                     type="submit"
                                     class="btn btn-info my-2"
                                     ${!resLocals.authz_result.authorized_edit ? 'disabled' : ''}
                                   >
-                                    Grade ${resLocals.savedAnswers} saved
-                                    ${resLocals.savedAnswers !== 1 ? 'answers' : 'answer'}
+                                    Grade ${savedAnswers} saved
+                                    ${savedAnswers !== 1 ? 'answers' : 'answer'}
                                   </button>
                                 `
                               : html`
@@ -416,17 +453,17 @@ export function StudentAssessmentInstance({
                                 `}
                           </form>
                           <ul class="my-1">
-                            ${resLocals.suspendedSavedAnswers > 1
+                            ${suspendedSavedAnswers > 1
                               ? html`
                                   <li>
-                                    There are ${resLocals.suspendedSavedAnswers} saved answers that
-                                    cannot be graded yet because their grade rate has not been
-                                    reached. They are marked with the
+                                    There are ${suspendedSavedAnswers} saved answers that cannot be
+                                    graded yet because their grade rate has not been reached. They
+                                    are marked with the
                                     <i class="fa fa-hourglass-half"></i> icon above. Reload this
                                     page to update this information.
                                   </li>
                                 `
-                              : resLocals.suspendedSavedAnswers === 1
+                              : suspendedSavedAnswers === 1
                                 ? html`
                                     <li>
                                       There is one saved answer that cannot be graded yet because
@@ -442,11 +479,10 @@ export function StudentAssessmentInstance({
                               the question page.
                             </li>
                             <li>
-                              Look at <strong>Submission status</strong> to confirm that each
-                              question has been graded. Questions with
-                              <strong>Available points</strong> can be attempted again for more
-                              points. Attempting questions again will never reduce the points you
-                              already have.
+                              Look at <strong>Status</strong> to confirm that each question has been
+                              graded. Questions with <strong>Available points</strong> can be
+                              attempted again for more points. Attempting questions again will never
+                              reduce the points you already have.
                             </li>
                             ${resLocals.authz_result.password != null ||
                             !resLocals.authz_result.show_closed_assessment
@@ -592,7 +628,7 @@ function InstanceQuestionTableHeader({ resLocals }: { resLocals: Record<string, 
             ? html`
                 <tr>
                   <th rowspan="2">Question</th>
-                  <th class="text-center" rowspan="2">Submission status</th>
+                  <th class="text-center" rowspan="2">Status</th>
                   <th class="text-center" colspan="2">Auto-grading</th>
                   <th class="text-center" rowspan="2">Manual grading points</th>
                   <th class="text-center" rowspan="2">Total points</th>
@@ -604,7 +640,7 @@ function InstanceQuestionTableHeader({ resLocals }: { resLocals: Record<string, 
             : html`
                 <tr>
                   <th>Question</th>
-                  <th class="text-center">Submission status</th>
+                  <th class="text-center">Status</th>
                   ${trailingColumns}
                 </tr>
               `}
@@ -632,23 +668,17 @@ function InstanceQuestionTableHeader({ resLocals }: { resLocals: Record<string, 
   `;
 }
 
-function ZoneInfoBadge({
-  popoverContent,
-  mainContent,
-}: {
-  popoverContent: string;
-  mainContent: string;
-}) {
+function ZoneInfoPopover({ label, content }: { label: string; content: string }) {
   return html`
     <button
       type="button"
-      class="btn btn-xs btn-secondary badge badge-secondary text-white font-weight-normal py-1"
+      class="btn btn-xs btn-secondary"
       data-toggle="popover"
       data-container="body"
       data-html="true"
-      data-content="${popoverContent}"
+      data-content="${content}"
     >
-      ${mainContent}&nbsp;<i class="far fa-question-circle" aria-hidden="true"></i>
+      ${label}&nbsp;<i class="far fa-question-circle" aria-hidden="true"></i>
     </button>
   `;
 }