+++ packages/formatter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/formatter
 
+## 1.3.4
+
+### Patch Changes
+
+- 4b79275: Upgrade all JavaScript dependencies
+
 ## 1.3.3
 
 ### Patch Changes